recommendation engine implemented with basic logic and recommended sessions are displayed on the map. Join session flow completed with bugs
