--- app/src/main/java/com/example/capstone/MapGestureHandler.java
@@ -1,7 +1,0 @@
-package com.example.capstone;
-
-import com.google.android.gms.maps.model.LatLng;
-
-public interface MapGestureHandler {
-    public void onZoomChange(int zoomLevel, LatLng currentLatLng);
-}
