--- app/src/main/java/com/example/capstone/NavigationProvider.java
@@ -1,5 +1,0 @@
-package com.example.capstone;
-
-public interface NavigationProvider {
-    public void navigate();
-}
