+++ app/src/main/java/com/example/capstone/activity/AuthenticationActivity.java
@@ -1,0 +1,37 @@
+package com.example.capstone.activity;
+
+import androidx.annotation.IdRes;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+
+import android.os.Bundle;
+
+import com.example.capstone.fragment.LoginFragment;
+import com.example.capstone.R;
+
+public class AuthenticationActivity extends AppCompatActivity {
+    public AuthenticationActivity() {
+        super(R.layout.activity_authentication);
+    }
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_authentication);
+        if (savedInstanceState == null) {
+            getSupportFragmentManager().beginTransaction()
+                    .setReorderingAllowed(true)
+                    .add(R.id.authentication, LoginFragment.class, null)
+                    .commit();
+        }
+    }
+
+    public void replaceFragment(@IdRes int containerViewId, @NonNull Class<? extends androidx.fragment.app.Fragment> fragmentClass){
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        fragmentManager.beginTransaction()
+                .replace(containerViewId, fragmentClass, null)
+                .setReorderingAllowed(true)
+                .addToBackStack(null)
+                .commit();
+    }
+}
