+++ app/src/main/java/com/example/capstone/fragment/SignupFragment.java
@@ -1,0 +1,124 @@
+package com.example.capstone.fragment;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import com.example.capstone.R;
+import com.example.capstone.model.User;
+import com.example.capstone.activity.AuthenticationActivity;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.parse.ParseException;
+import com.parse.SaveCallback;
+
+public class SignupFragment extends Fragment {
+    private static final String TAG = "SignupFragment";
+    private FirebaseAuth mAuth;
+
+    private Button btnSignupUser;
+    private EditText etSignupEmail;
+    private EditText etSignupPassword;
+
+    public SignupFragment() {}
+
+    public static SignupFragment newInstance(String param1, String param2) {
+        SignupFragment fragment = new SignupFragment();
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        mAuth = FirebaseAuth.getInstance();
+    }
+
+    private void saveParseUser(FirebaseUser firebaseUser){
+        User parseUser = new User();
+        parseUser.setFirebaseUid(firebaseUser.getUid());
+        parseUser.saveInBackground(new SaveCallback() {
+            @Override
+            public void done(ParseException e) {
+                if (e != null){
+                    Log.e(TAG, "Error while saving", e);
+                    Toast.makeText(getActivity(), "Error while saving!", Toast.LENGTH_SHORT).show();
+                }
+                Log.i(TAG, "Registration was successful!");
+            }
+        });
+    }
+
+    private void createAccount(String email, String password) {
+        mAuth.createUserWithEmailAndPassword(email, password)
+                .addOnCompleteListener(getActivity(), new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            Log.d(TAG, "createUserWithEmail:success");
+                            FirebaseUser firebaseUser = mAuth.getCurrentUser();
+                            saveParseUser(firebaseUser);
+                            updateUI(firebaseUser);
+                        } else {
+                            Log.w(TAG, "createUserWithEmail:failure", task.getException());
+                            Toast.makeText(getActivity(), "Authentication failed.", Toast.LENGTH_SHORT).show();
+                            updateUI(null);
+                        }
+                    }
+                });
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        // Check if user is signed in (non-null) and update UI accordingly.
+        FirebaseUser currentUser = mAuth.getCurrentUser();
+        if(currentUser != null){
+            reload();
+        }
+    }
+
+    private void reload() { }
+
+    private void updateUI(FirebaseUser user) {}
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_signup, container, false);
+    }
+
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+        etSignupEmail = getActivity().findViewById(R.id.et_signup_email);
+        etSignupPassword = getActivity().findViewById(R.id.et_signup_password);
+        btnSignupUser = ((AuthenticationActivity)getActivity()).findViewById(R.id.btn_signup_user);
+        btnSignupUser.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Log.i(TAG, "onClick signup button");
+                String email = etSignupEmail.getText().toString();
+                String password = etSignupPassword.getText().toString();
+
+                if (email.matches("") || password.matches("")) {
+                    Toast.makeText(getActivity(), "Please enter your email and password.", Toast.LENGTH_SHORT).show();
+                } else {
+                    createAccount(email, password);
+                    ((AuthenticationActivity)getActivity()).replaceFragment(R.id.authentication, ProfileCreationFragment.class);
+                }
+            }
+        });
+    }
+}
