+++ app/src/main/java/com/example/capstone/utils/MapGestureHandler.java
@@ -1,0 +1,7 @@
+package com.example.capstone.utils;
+
+import com.google.android.gms.maps.model.LatLng;
+
+public interface MapGestureHandler {
+    public void onZoomChange(int zoomLevel, LatLng currentLatLng);
+}
