+++ app/src/main/java/com/example/capstone/utils/NavigationProvider.java
@@ -1,0 +1,5 @@
+package com.example.capstone.utils;
+
+public interface NavigationProvider {
+    public void navigate();
+}
