+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:id="@+id/homescreen"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomescreenActivity">
+    tools:context=".activity.HomescreenActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_joined_study_sessions"
