+++ app/src/main/res/layout/fragment_join_session.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".JoinSessionFragment">
+    tools:context=".fragment.JoinSessionFragment">
 
 <!--    <CheckBox-->
 <!--        android:id="@+id/checkBox3"-->
