+++ app/src/main/res/layout/fragment_profile_creation.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".ProfileCreationFragment">
+    tools:context=".fragment.ProfileCreationFragment">
 
     <ImageView
         android:id="@+id/iv_profile_picture"
