+++ app/src/main/res/layout/fragment_signup.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".SignupFragment">
+    tools:context=".fragment.SignupFragment">
 
     <ImageView
         android:id="@+id/logo_signup"
