+++ app/src/main/res/layout/fragment_start_session_logistics.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".StartSessionLogisticsFragment">
+    tools:context=".fragment.StartSessionLogisticsFragment">
 
     <EditText
         android:id="@+id/et_start_session_name"
