+++ app/src/main/res/layout/fragment_start_session_map.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".StartSessionMapFragment">
+    tools:context=".fragment.StartSessionMapFragment">
 
     <fragment xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:map="http://schemas.android.com/apk/res-auto"
