+++ app/src/main/res/layout/fragment_start_session_preferences.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".StartSessionPreferencesFragment">
+    tools:context=".fragment.StartSessionPreferencesFragment">
 
     <EditText
         android:id="@+id/et_start_session_subjects"
